Cleaned the project directories and lines fo code
