--- thank_you.php
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Thank You</title>
-</head>
-<body>
-    <h1>Thank You for Your Feedback!</h1>
-    <p>We appreciate you taking the time to provide feedback. Your input helps us improve.</p>
-    <!-- You can add additional content or links here -->
-</body>
-</html>
